==> Apriori_new_patient_data/1-MergeTranslated&NewStats.ipynb <==
import pandas as pd
import numpy as np

pd.set_option("display.max_columns", None)

# from mlxtend.frequent_patterns import apriori, association_rules

# Load the data
data = pd.read_csv("data/translated.csv")
display("Columns before dropping :",data.columns)
drop_cols = [
    # "occupation_en",
    # "address_en",
    "M_mean",
    "M_median",
    "M_std",
    "M_min",
    "M_max",
    "A_mean",
    "A_median",
    "A_std",
    "A_min",
    "A_max",
    "20th_quantile",
    "25th_quantile",
    "50th_quantile",
    "75th_quantile",
]  # columns to drop
data.drop(drop_cols, axis=1, inplace=True)
display("Columns after dropping :",data.columns)
display(data)

# Load the data
data_stats = pd.read_csv("data/patient_statistics_new.csv")

display(data_stats)

data = data.merge(data_stats, left_on="ID", right_on="Patient_ID")

data.drop(["Patient_ID"], axis=1, inplace=True)

# Rename the first column to "Patient_ID"
data = data.rename(columns={data.columns[0]: "Patient_ID"})

display(data)

del data_stats

data.to_csv("data/translated_new.csv", index=False)

del data

data = pd.read_csv("data/translated_new.csv")

drop_cols = [
    "occupation_en",
    "address_en",
]  # columns to drop

data.drop(drop_cols, axis=1, inplace=True)
display(data)

del drop_cols
