==> Apriori_new_patient_data/2-BinnedColumns_DropNan.ipynb <==
import pandas as pd
import numpy as np

pd.set_option("display.max_columns", None)

data = pd.read_csv("data/translated_new.csv")

drop_cols = [
    "occupation_en",
    "address_en",
    "occupation_label"
]  # columns to drop

data.drop(drop_cols, axis=1, inplace=True)
display(data)

del drop_cols

Sex = ["M", "F"]

Smoke = ["NS", "ES", "SM"]

occupation_category = [
    "Housewife",
    "Manufacturing",
    "Construction",
    "Other",
    "Office",
    "Education",
    "Finance",
    "Unemployed",
    "Service Industry",
    "Agriculture",
    "Healthcare",
    "Transportation",
]
# occupation_label = [5, 6, 1, 8, 7, 2, 3, 11, 9, 0, 4, 10]


# Occupation = {
#     "Housewife": 5,
#     "Manufacturing": 6,
#     "Construction": 1,
#     "Other": 8,
#     "Office": 7,
#     "Education": 2,
#     "Finance": 3,
#     "Unemployed": 11,
#     "Service Industry": 9,
#     "Agriculture": 0,
#     "Healthcare": 4,
#     "Transportation": 10,
# }

import pandas as pd
import numpy as np


def bin_column(df, column_name, num_bins=4):
    # Create the bins
    bins = pd.qcut(df[column_name].dropna(), q=num_bins, labels=False)

    # Create labels for the bins
    labels = [f"{column_name}_Q{i+1}" for i in range(num_bins)]

    # Add the binned column to the dataframe
    df[f"{column_name}_binned"] = pd.cut(df[column_name], bins=num_bins, labels=labels)

    return df


cols_for_binning = [
    "A_20th_quantile",
    "A_25th_quantile",
    "A_50th_quantile",
    "A_75th_quantile",
    "A_max",
    "A_mean",
    "A_median",
    "A_min",
    "A_std",
    "M_20th_quantile",
    "M_25th_quantile",
    "M_50th_quantile",
    "M_75th_quantile",
    "M_max",
    "M_mean",
    "M_median",
    "M_min",
    "M_std",
]

for col in cols_for_binning:
    data = bin_column(data, col)


data.drop(cols_for_binning, axis=1, inplace=True)

del cols_for_binning

# Now df contains new columns with the binned data

data.dropna(inplace=True)

print(f"data.columns = {list(data.columns)}")

data.to_csv("data/translated_new_binned_DropNaN.csv", index=False)

data = pd.read_csv("data/translated_new_binned_DropNaN.csv")

display(data)
